Refactored Commands into packages
