+++ src/main/java/optix/commands/seats/SellSeatCommand.java
@@ -1,6 +1,7 @@
-package optix.commands;
+package optix.commands.seats;
 
 import optix.Ui;
+import optix.commands.Command;
 import optix.constant.OptixResponse;
 import optix.core.Storage;
 import optix.core.Theatre;
