+++ src/main/java/optix/commands/shows/AddCommand.java
@@ -1,6 +1,7 @@
-package optix.commands;
+package optix.commands.shows;
 
 import optix.Ui;
+import optix.commands.Command;
 import optix.constant.OptixResponse;
 import optix.core.Storage;
 import optix.core.Theatre;
@@ -35,7 +36,6 @@ public class AddCommand extends Command {
         if (showLocalDate.compareTo(today) <= 0) {
             ui.setMessage("☹ OOPS!!! It is not possible to perform in the past.\n");
         } else if (shows.containsKey(showLocalDate)) {
-            // to abstract the  date formatting in the future extensions
             ui.setMessage("☹ OOPS!!! There is already a show being added on that date.\n"
                     + "Please try again. \n");
         } else {
