+++ src/test/java/Optix/commands/AddCommandTest.java
@@ -1,7 +1,7 @@
 package Optix.commands;
 
 import optix.Ui;
-import optix.commands.AddCommand;
+import optix.commands.shows.AddCommand;
 import optix.core.Storage;
 import optix.util.ShowMap;
 import org.junit.jupiter.api.Test;
