+++ src/test/java/Optix/commands/ByeCommandTest.java
@@ -1,7 +1,6 @@
 package Optix.commands;
 
 import optix.Ui;
-import optix.commands.AddCommand;
 import optix.commands.ByeCommand;
 import optix.core.Storage;
 import optix.util.ShowMap;
