+++ src/test/java/Optix/commands/DeleteOneCommandTest.java
@@ -1,8 +1,8 @@
 package Optix.commands;
 
 import optix.Ui;
-import optix.commands.AddCommand;
-import optix.commands.DeleteOneCommand;
+import optix.commands.shows.AddCommand;
+import optix.commands.shows.DeleteOneCommand;
 import optix.core.Storage;
 import optix.util.ShowMap;
 import org.junit.jupiter.api.Test;
